refactor: extract static data from tool pages
Extracted large constants, types, and configurations from `app/tools` component files into dedicated `-data.ts` files to reduce file size and maintain separation of concerns. Affected tools: `aws-diagram`, `invisible-characters`, `lorem-ipsum`, and `favicon-generator`.
