+++ app/tools/lorem-ipsum/lorem-ipsum-data.ts
@@ -1,0 +1,123 @@
+export const LOREM_TEXTS = {
+  lorem: {
+    name: 'Lorem Ipsum (ラテン語)',
+    paragraphs: [
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
+      'Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
+      'Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo.',
+      'Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet.',
+      'Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae.',
+    ],
+    words: [
+      'lorem',
+      'ipsum',
+      'dolor',
+      'sit',
+      'amet',
+      'consectetur',
+      'adipiscing',
+      'elit',
+      'sed',
+      'do',
+      'eiusmod',
+      'tempor',
+      'incididunt',
+      'ut',
+      'labore',
+      'et',
+      'dolore',
+      'magna',
+      'aliqua',
+      'enim',
+      'ad',
+      'minim',
+      'veniam',
+      'quis',
+      'nostrud',
+      'exercitation',
+      'ullamco',
+      'laboris',
+      'nisi',
+      'aliquip',
+      'ex',
+      'ea',
+      'commodo',
+      'consequat',
+      'duis',
+      'aute',
+      'irure',
+      'reprehenderit',
+      'voluptate',
+      'velit',
+      'esse',
+      'cillum',
+      'fugiat',
+      'nulla',
+      'pariatur',
+      'excepteur',
+      'sint',
+      'occaecat',
+      'cupidatat',
+      'non',
+      'proident',
+      'sunt',
+      'culpa',
+      'qui',
+      'officia',
+      'deserunt',
+      'mollit',
+      'anim',
+      'id',
+      'est',
+      'laborum',
+    ],
+  },
+  neko: {
+    name: '吾輩は猫である (夏目漱石)',
+    paragraphs: [
+      '吾輩は猫である。名前はまだ無い。どこで生れたかとんと見当がつかぬ。何でも薄暗いじめじめした所でニャーニャー泣いていた事だけは記憶している。吾輩はここで始めて人間というものを見た。しかもあとで聞くとそれは書生という人間中で一番獰悪な種族であったそうだ。この書生というのは時々我々を捕えて煮て食うという話である。',
+      'しかしその当時は何という考もなかったから別段恐ろしいとも思わなかった。ただ彼の手のひらに載せられてスーと持ち上げられた時、何だかフワフワした感じがあったばかりである。書生の顔の上一つ乗っているのが、いささか心細いばかりであった。',
+      'おまけに吾輩の顔のまん中へ、何だか暖かいような、それでいてすこぶるざらざらしたものが時々通り過ぎる。これが書生が飯を食ったあとに、茶碗の底に残った滓を舐める時の舌だとは、のちになってようやく知ったのである。',
+      'そのころはただ不思議なものだと思って、別段深い詮索もしなかったが、その後だんだん知恵がついてみると、人間ほど不人情な動物はいないと思うようになった。第一我々の同族を捕えて皮を剥いだり肉を食ったり、それをおのれらの衣食住の用にあてるなどとは言語道断の沙汰である。',
+    ],
+    sentences: [
+      '吾輩は猫である。',
+      '名前はまだ無い。',
+      'どこで生れたかとんと見当がつかぬ。',
+      '何でも薄暗いじめじめした所でニャーニャー泣いていた事だけは記憶している。',
+      '吾輩はここで始めて人間というものを見た。',
+      'しかもあとで聞くとそれは書生という人間中で一番獰悪な種族であったそうだ。',
+      'この書生というのは時々我々を捕えて煮て食うという話である。',
+      'しかしその当時は何という考もなかった。',
+      'ただ彼の手のひらに載せられてスーと持ち上げられた時、フワフワした感じがあった。',
+    ],
+  },
+  melos: {
+    name: '走れメロス (太宰治)',
+    paragraphs: [
+      'メロスは激怒した。必ず、かの邪智暴虐の王を除かなければならぬと決意した。メロスには政治がわからぬ。メロスは、村の牧人である。笛を吹き、羊と遊んで暮して来た。けれども邪悪に対しては、人一倍に敏感であった。きょう未明メロスは村を出発し、野を越え山越え、十里はなれた此のシラクスの市にやって来た。',
+      'メロスには無二の友があった。セリヌンティウスである。今は此のシラクスの市で、石工をしている。その友を、これから訪ねてみるつもりなのだ。久しく逢わなかったのだから、訪ねて行くのが楽しみである。歩きながらメロスは、市の中の様子に、何か釈然としないものを感じた。',
+      'ひっそりしている。もう既に日も落ちて、街の暗いのは当りまえだが、けれども、なんだか、夜のせいばかりでは無く、市全体が、やけに寂しい。のんきなメロスも、だんだん不安になって来た。路で逢った若い衆に、何かあったのかと問いかけた。王は、人を殺します。なぜ殺すのだ。悪心を抱いている、というのでございますが、誰もそんな、悪心を持っていはいたしませぬ。',
+    ],
+    sentences: [
+      'メロスは激怒した。',
+      '必ず、かの邪智暴虐の王を除かなければならぬと決意した。',
+      'メロスには政治がわからぬ。',
+      'メロスは、村の牧人である。',
+      '笛を吹き、羊と遊んで暮して来た。',
+      'けれども邪悪に対しては、人一倍に敏感であった。',
+      'メロスには無二の友があった。',
+      'セリヌンティウスである。',
+      '久しく逢わなかったのだから、訪ねて行くのが楽しみである。',
+    ],
+  },
+};
+
+export type TextType = keyof typeof LOREM_TEXTS;
+export type UnitType = 'paragraphs' | 'sentences' | 'words' | 'list';
+
+// 決定論的な疑似乱数生成器 (シード値に基づく)
+export function pseudoRandom(seed: number) {
+  const x = Math.sin(seed) * 10000;
+  return x - Math.floor(x);
+}
